+++ NEWS.md
@@ -18,6 +18,7 @@
 * `lines_of_r_code()` now works correctly reading a single file
 * `import(, overwrite = TRUE)` now works
 * `ls_function()`, `ls_object()`, `ls_dataframe()`, and `ls_all()` have improvements for environmental searching
+* `assign_labels.data.frame(.missing = "warning")` correctly removes missing columns
 
 ## Non visible changes